feat: enable low-risk linters (unconvert, wastedassign, usestdlibvars) (#23)
* feat: enable low-risk linters (unconvert, wastedassign, usestdlibvars)

Phase 1 of linting improvements - enable linters that catch real bugs
and improve code quality with minimal false positives.

Changes:
- Enable unconvert: Remove unnecessary type conversions (123 fixes)
- Enable wastedassign: Fix unused variable assignments (3 fixes)
- Enable usestdlibvars: Use stdlib constants (98 fixes)
  - HTTP status codes: use http.StatusOK, http.StatusNotFound, etc.
  - HTTP methods: use http.MethodGet, http.MethodPost, etc.
  - Time layouts: use time.DateTime, time.DateOnly, time.TimeOnly

Total: 224 linting issues resolved across 72 files

All tests pass. No regressions introduced.

* fix: use copy() instead of loop for slice assignment (staticcheck S1001)

+++ internal/adapters/v0_0_42/account_adapter.go
@@ -285,7 +285,7 @@ func (a *AccountAdapter) convertAccountAssociationRequestToAPI(req *types.Accoun
 
 		// Set fairshare (note: field name is "Fairshare" in v0.0.42)
 		if req.Fairshare != 0 {
-			fairshareInt32 := int32(req.Fairshare)
+			fairshareInt32 := req.Fairshare
 			assocRec.Fairshare = &fairshareInt32
 		}
