+++ internal/adapters/v0_0_42/standalone_adapter.go
@@ -420,7 +420,7 @@ func (a *StandaloneAdapter) GetTRES(ctx context.Context) (*types.TRESList, error
 			tres.Name = *apiTres.Name
 		}
 		if apiTres.Count != nil {
-			tres.Count = int64(*apiTres.Count)
+			tres.Count = *apiTres.Count
 		}
 
 		tresList = append(tresList, tres)
@@ -449,7 +449,7 @@ func (a *StandaloneAdapter) CreateTRES(ctx context.Context, req *types.CreateTRE
 	}
 
 	if req.Count > 0 {
-		count := int64(req.Count)
+		count := req.Count
 		apiReq.TRES[0].Count = &count
 	}
 
@@ -473,7 +473,7 @@ func (a *StandaloneAdapter) CreateTRES(ctx context.Context, req *types.CreateTRE
 	}
 
 	if req.Count > 0 {
-		tres.Count = int64(req.Count)
+		tres.Count = req.Count
 	}
 
 	return tres, nil
